// requests-service/src/infrastructure/controllers/requests/dtos/offerDetails.dto.ts
export class OfferDetailsDTO {
  offerId: string;
  offerName: string;
  offerDescription: string;
  offerImage: string;
  offerProvider: string;
  activeStatus: string;
  tenure: number;
  roi: string;
  noOfInstallment: string;
  repaymentFrequency: string;
  offerLimit: number;
  applicationFee: string;
}